requirements.txt
attrs==23.2.0
beautifulsoup4==4.12.3
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
click-plugins==1.1.1
cligj==0.7.2
easyocr==1.7.1
ExifRead==3.0.0
filelock==3.13.4
fiona==1.9.6
fsspec==2024.3.1
geopandas==0.14.4
gpsphoto==2.2.3
idna==3.7
imageio==2.34.1
img2table==1.2.11
Jinja2==3.1.3
lazy_loader==0.4
llvmlite==0.42.0
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.3
ninja==1.11.1.1
numba==0.59.1
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.1.105
opencv-contrib-python==4.9.0.80
opencv-python-headless==4.9.0.80
packaging==24.0
pandas==2.2.2
pillow==10.3.0
polars==0.20.22
pyarrow==16.0.0
pyclipper==1.3.0.post5
PyMuPDF==1.24.2
PyMuPDFb==1.24.1
pyproj==3.6.1
pytesseract==0.3.10
python-bidi==0.4.2
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
scikit-image==0.23.2
scipy==1.13.0
shapely==2.0.4
six==1.16.0
soupsieve==2.5
sympy==1.12
tesseract==0.1.3
tifffile==2024.4.18
torch==2.2.2
torchvision==0.17.2
tqdm==4.66.2
triton==2.2.0
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
XlsxWriter==3.2.0
